Move i2c_lcd tests to separate file
