+++ lcd/i2c_lcd.py
@@ -87,40 +87,3 @@ class I2cLcd(LcdApi):
                 ((data & 0x0f) << SHIFT_DATA))
         self.bus.write_byte(self.i2c_addr, byte | MASK_E)
         self.bus.write_byte(self.i2c_addr, byte)
-
-def test_main():
-    """Test function for verifying basic functionality."""
-    lcd = I2cLcd(1, 0x27, 2, 16)
-    lcd.putstr("It Works!\nSecond Line")
-    time.sleep(3)
-    lcd.clear()
-    count = 0
-    while True:
-        lcd.move_to(0, 0)
-        lcd.putstr(time.strftime('%b %d %Y\n%H:%M:%S', time.localtime()))
-        time.sleep(1)
-        count += 1
-        if count % 10 == 3:
-            print "Turning backlight off"
-            lcd.backlight_off()
-        if count % 10 == 4:
-            print "Turning backlight on"
-            lcd.backlight_on()
-        if count % 10 == 5:
-            print "Turning display off"
-            lcd.display_off()
-        if count % 10 == 6:
-            print "Turning display on"
-            lcd.display_on()
-        if count % 10 == 7:
-            print "Turning display & backlight off"
-            lcd.backlight_off()
-            lcd.display_off()
-        if count % 10 == 8:
-            print "Turning display & backlight on"
-            lcd.backlight_on()
-            lcd.display_on()
-
-if __name__ == "__main__":
-    test_main()
-
